requirements.txt
Flask==0.10.1
Flask-Bootstrap==3.3.5.3
Flask-Login==0.2.11
Flask-Moment==0.5.0
Flask-Script==2.0.5
Flask-WTF==0.9.5
pillow==2.9.0
requests==2.7.0